refactor endpoint rule path matcher
